+++ python/transitio/gtfs/__init__.py
@@ -1,7 +1,13 @@
 """GTFS transformations: feed cropping, merging and comparison."""
 
-from transitio.gtfs._compare import compare_feeds
+from transitio.gtfs._compare import compare_feed_history, compare_feeds
 from transitio.gtfs._crop import crop_feed
 from transitio.gtfs._merge import merge_feeds, merge_tables
 
-__all__ = ["compare_feeds", "crop_feed", "merge_feeds", "merge_tables"]
+__all__ = [
+    "compare_feed_history",
+    "compare_feeds",
+    "crop_feed",
+    "merge_feeds",
+    "merge_tables",
+]